Added two data files
newPostmasters: appointments on establishment dates are marked -1,
other appointments 0, bad data entries -1
newPostmastersmulti: appointments on establishment dates are marked +1
to +8.In particular, if the establishment date was recorded in Est4,
the number is +4. Similarly, appointments on name change dates are
marked with negative numbers. Bad entries are -50 and regular
appointments are 0.
The labels are under the field NewApp
Out of 38525 post masters, about 8938 were appointment during
establishment/name changes i.e. a ratio of about 23%, which is quite
high.